add maze-generator dependency and parcel plugin for resolving a depedency
Installing the maze-generator dependency install another dependency named "prando", which has a problem being imported by the maze-generator dependency. I think it's related to the maze-generator package being a commonjs node module and it imports "prando" has a ES6 module, potentially while being bundled by parcel (?). So a plugin for parcel was created to re-route the import statement to the commonjs version of "prando".